feat(cosmosdb): add Azure Cosmos DB Table API support (M6)
Cosmos's Table API is wire-identical to Table Storage's OData/JSON
entity model, differing only in auth and endpoint conventions. Extract
the shared entity CRUD/$filter engine out of services/azuretable into
pkgs/odatatable/ so both services import the same logic, then route
Cosmos's existing Gateway port (8081) to it by path shape, reusing
Cosmos's existing master-key auth unchanged. $batch stays deferred,
matching Table Storage's own M2 deferral.
