+++ README.md
@@ -690,7 +690,7 @@ Every service links to its own page with a coverage breakdown — audited operat
 | Service | Parity | PARITY Entries | Notes |
 |---|---|---|---|
 | [Azure Blob Storage](services/azureblob/README.md) | C | 8 | 8 gaps; 2 deferred |
-| [Azure Cosmos DB](services/cosmosdb/README.md) | C | 15 | 9 gaps; 5 deferred |
+| [Azure Cosmos DB](services/cosmosdb/README.md) | C | 15 | 10 gaps; 6 deferred |
 | [Azure Queue Storage](services/azurequeue/README.md) | C | 9 | 7 gaps; 1 deferred |
 | [Azure Service Bus](services/azureservicebus/README.md) | B | 17 | 4 gaps; 2 deferred |
 | [Azure Table Storage](services/azuretable/README.md) | C | 10 | 6 gaps; 3 deferred |
