+++ services/azuretable/PARITY.md
@@ -20,9 +20,9 @@ ops:
 families:
   auth: {status: partial, note: "Identical stance to services/azurequeue: checkAuth parses a present Authorization header via pkgs/azureauth.ParseAuthorizationHeader (structural only, SharedKey and SharedKeyLite both accepted). Verification is not enforced; an absent or malformed header is still accepted."}
   wire_protocol: {status: ok, note: "REST+JSON/OData. Honors the Accept header's odata= level (nometadata/minimalmetadata/fullmetadata), responding with a matching Content-Type. x-ms-version (pinned to 2019-02-02, the literal value azure-sdk-for-go/sdk/data/aztables' generated client sends and expects), x-ms-request-id, Date, and DataServiceVersion: 3.0; are set on every response; x-ms-error-code is set on every error."}
-  odata_filter: {status: ok, note: "Hand-written lexer -> recursive-descent parser -> AST -> evaluator (odata_filter.go/odata_filter_eval.go), modeled on services/dynamodb/expr. Supports eq/ne/lt/le/gt/ge, and/or/not, parentheses, and every literal form (quoted string with '' escape, integer, Int64 'L' suffix, float, true/false, datetime'..', guid'..', X'..'/binary'..'). Recursion is depth-bounded (maxFilterDepth=100), counted per genuine nesting level ('not' and '(...)' only, not per precedence-climbing layer -- see Notes) against stack-overflow DoS. A comparison against a missing property evaluates false, never an error; a parse error always surfaces as 400 InvalidInput, never a panic. Two Int64 operands are compared as int64 directly, not via a blanket float64 conversion, so magnitudes beyond 2^53 compare correctly (see Notes)."}
+  odata_filter: {status: ok, note: "Hand-written lexer -> recursive-descent parser -> AST -> evaluator, extracted into pkgs/odatatable (filter.go/eval.go) per AZURE.md section 9's M6 milestone so services/cosmosdb's Table API could share it; re-exported here under this package's historical names (see interfaces.go). Modeled on services/dynamodb/expr. Supports eq/ne/lt/le/gt/ge, and/or/not, parentheses, and every literal form (quoted string with '' escape, integer, Int64 'L' suffix, float, true/false, datetime'..', guid'..', X'..'/binary'..'). Recursion is depth-bounded (maxFilterDepth=100), counted per genuine nesting level ('not' and '(...)' only, not per precedence-climbing layer -- see Notes) against stack-overflow DoS. A comparison against a missing property evaluates false, never an error; a parse error always surfaces as 400 InvalidInput, never a panic. Two Int64 operands are compared as int64 directly, not via a blanket float64 conversion, so magnitudes beyond 2^53 compare correctly (see Notes)."}
   edm_types: {status: ok, note: "Edm.String/Int32/Int64/Double/Boolean/DateTime/Guid/Binary all round-trip. Int64/DateTime/Guid/Binary emit a Prop@odata.type annotation on read (matching real Table Storage and aztables' own EDMEntity.MarshalJSON); Int32/Double/Boolean/String do not. Unannotated bare-number decode tries Int32 first, then falls back to Double, exactly mirroring aztables' own client-side inference -- a whole-number Double (e.g. 4.0) is therefore indistinguishable from an Int32 without an explicit annotation, an inherent wire-protocol ambiguity aztables itself has (its own MarshalJSON never annotates plain float64 either). EdmBinary values are deep-copied on every store/return path (insert, merge, get, query) so a caller mutating a []byte it passed in or received back can never alias/corrupt stored state."}
-  etag_and_timestamp: {status: ok, note: "ETag format W/\"datetime'<url-encoded RFC3339, 7 fractional digits>'\" derived from Timestamp. Every mutation advances Timestamp by at least 100ns past its previous value even if the injected clock returns the same instant twice in a row, guaranteeing a distinct ETag on every write (see store.go's bumpTimestamp) -- covered by TestInMemoryBackend_ETagChangesOnEveryWrite."}
+  etag_and_timestamp: {status: ok, note: "ETag format W/\"datetime'<url-encoded RFC3339, 7 fractional digits>'\" derived from Timestamp. Every mutation advances Timestamp by at least 100ns past its previous value even if the injected clock returns the same instant twice in a row, guaranteeing a distinct ETag on every write (see pkgs/odatatable/store.go's bumpTimestamp, as of the M6 extraction) -- covered by TestInMemoryBackend_ETagChangesOnEveryWrite."}
   persistence: {status: ok, note: "Snapshot version 2 (bumped from 1 -- see Notes). Edm.Int64 is snapshotted as a decimal string, not a bare JSON number: a float64 number loses precision above 2^53, corrupting large Int64 values across a save/restore cycle. Entity identity within a table is keyed by entityCompositeKey (a {PartitionKey, RowKey} struct, persisted via MarshalText as a JSON string array), not a NUL-delimited string, closing a key-collision class where two different (PartitionKey, RowKey) pairs could hash to the same delimited string. Restore initializes a table's Entities map when the snapshot carries \"Entities\": null (legal JSON, decodes to a nil map) rather than leaving it nil, which would panic the first time anything inserts into it."}
   routing_isolation: {status: ok, note: "Runs on its own dedicated *http.Server, bound synchronously in StartWorker to a fixed port (default 10002 via --azure-table-port/AZURE_TABLE_PORT, matching Azurite's own Table service port; no fallback pool -- fails fast if unavailable, mirroring services/azureblob and services/azurequeue). cli.go's reserveFixedServicePorts additionally reserves this port in the shared PortAlloc pool since 10002 sits inside --port-range-start/--port-range-end's own default range."}
   observability: {status: ok, note: "StartWorker wraps its Echo handler with telemetry.WrapEchoHandler so ExtractOperation/ExtractResource feed Prometheus metrics. InMemoryBackend uses *lockmetrics.RWMutex instead of raw sync.RWMutex, matching repo convention."}
@@ -75,9 +75,14 @@ operand    := identifier | literal
 literal    := 'quoted string' (with '' escape) | integer | integer'L' | float | true | false
             | datetime'<RFC3339>' | guid'<uuid>' | X'<hex>' | binary'<base64>'
 ```
-Implemented as a real lexer -> recursive-descent parser -> AST -> evaluator
-(`odata_filter.go`/`odata_filter_eval.go`), modeled on
-`services/dynamodb/expr`'s identical shape -- not string matching. Recursion
+Implemented as a real lexer -> recursive-descent parser -> AST -> evaluator,
+now living in `pkgs/odatatable` (`filter.go`/`eval.go`) since AZURE.md
+section 9's M6 milestone extracted it out of this package so
+`services/cosmosdb`'s Table API could share the same engine; this package
+re-exports `Node`/`ParseFilter`/`EvaluateFilter` under their historical
+names (see `interfaces.go`/`odata_filter.go`/`odata_filter_eval.go`).
+Modeled on `services/dynamodb/expr`'s identical shape -- not string
+matching. Recursion
 depth is bounded (`maxFilterDepth = 100`) so a maliciously (or accidentally)
 deeply-nested filter fails with a parse error instead of overflowing the
 stack. Identifiers resolve against `PartitionKey`/`RowKey`/`Timestamp` first,
