+++ services/azuretable/interfaces.go
@@ -5,11 +5,22 @@
 // azure-sdk-for-go clients to operate against. See AZURE.md and PARITY.md
 // for scope and known gaps.
 //
+// The entity CRUD/$filter engine itself lives in pkgs/odatatable (extracted
+// here in AZURE.md section 9's M6 milestone so services/cosmosdb's Table API
+// could import the same engine instead of duplicating it); this package is a
+// thin wire-protocol adapter over it -- HTTP routing/dispatch (handler.go),
+// request/response shaping (table_ops.go/entity_ops.go), and re-exports of
+// pkgs/odatatable's types/functions/errors under their historical
+// azuretable-local names (models.go, errors.go, store.go) so this package's
+// own public API is unaffected by the extraction.
+//
 // Unlike services/azurequeue and services/azureblob, this package has no
 // janitor.go: Table Storage entities carry no TTL/expiry concept (there is
 // no message-visibility or blob-lease analogue), so there is nothing for a
 // background sweep to do.
 package azuretable
+
+import "github.com/blackbirdworks/gopherstack/pkgs/odatatable"
 
 // Compile-time assertion: InMemoryBackend must implement StorageBackend.
 var _ StorageBackend = (*InMemoryBackend)(nil)
@@ -18,12 +29,12 @@ var _ StorageBackend = (*InMemoryBackend)(nil)
 // exist, but match unconditionally on ETag" -- as opposed to an empty
 // ifMatch (no header at all, meaning upsert semantics) or a specific ETag
 // string (optimistic-concurrency match required).
-const IfMatchAny = "*"
+const IfMatchAny = odatatable.IfMatchAny
 
 // StorageBackend defines the interface for an Azure Table Storage backend.
-// Shaped after services/azurequeue's StorageBackend: a narrow, testable seam
-// between the wire handler and storage, so handler tests can substitute a
-// fake.
+// Re-exported from pkgs/odatatable.StorageBackend (see this file's package
+// doc comment) so existing callers referencing azuretable.StorageBackend
+// keep compiling unchanged.
 //
 // The ifMatch parameter on ReplaceEntity/MergeEntity/DeleteEntity threads
 // through the three If-Match states the wire protocol distinguishes:
@@ -34,26 +45,4 @@ const IfMatchAny = "*"
 //   - IfMatchAny ("*"): the entity must exist, but any current ETag matches.
 //   - any other string: the entity must exist AND its current ETag must
 //     equal this value, else ErrETagMismatch.
-type StorageBackend interface {
-	CreateTable(name string) error
-	DeleteTable(name string) error
-	ListTables() []TableInfo
-
-	InsertEntity(table, partitionKey, rowKey string, props map[string]EntityProperty) (EntityInfo, error)
-	GetEntity(table, partitionKey, rowKey string) (EntityInfo, error)
-	// QueryEntities returns entities in table matching filter (nil matches
-	// everything), ordered by (PartitionKey, RowKey), capped at top results
-	// (top <= 0 means unlimited).
-	QueryEntities(table string, filter Node, top int) ([]EntityInfo, error)
-	ReplaceEntity(
-		table, partitionKey, rowKey string,
-		props map[string]EntityProperty,
-		ifMatch string,
-	) (EntityInfo, error)
-	MergeEntity(table, partitionKey, rowKey string, props map[string]EntityProperty, ifMatch string) (EntityInfo, error)
-	DeleteEntity(table, partitionKey, rowKey, ifMatch string) error
-
-	// Reset clears all in-memory state. Used by the
-	// POST /_gopherstack/reset endpoint for CI pipelines and rapid local development.
-	Reset()
-}
+type StorageBackend = odatatable.StorageBackend
